Move flaky test out to the `syncIntegrationTest` source set
